pyproject.toml
[project]
name = "thesis-metrics"
version = "0.1.0"
description = "Privacy metrics evaluation for synthetic medical data"
authors = [
    {name = "Simon", email = "simonmeoni@aol.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "hydra-core>=1.3.0",
    "rich>=13.0.0",
    "plotext>=5.2.0",
    "wandb>=0.16.0",
    "sentence-transformers>=5.1.2",
    "groq>=0.4.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
alpacare = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "trl>=0.7.0",
    "peft>=0.5.0",
    "accelerate>=0.20.0",
    "vllm>=0.2.0",
    "openai>=1.0.0",
]

downstream = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "evaluate>=0.4.0",
    "seqeval>=1.2.2",
    "nltk>=3.8.0",
]

dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipython>=8.15.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[project.scripts]
privacy-metrics = "thesis_metrics.cli.privacy_metrics:main"
rephrase = "thesis_metrics.cli.rephrase:main"
alpacare-eval = "thesis_metrics.cli.alpacare_eval:main"
downstream-eval = "thesis_metrics.cli.downstream_eval:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["thesis_metrics*"]
exclude = ["data*", "slurm*", "outputs*", "wandb*", "dstream_tasks*"]
